Untangle errors enum and update documentation (#105)
* untangle errors

* update documentation

* update readme
