+++ dsh_sdk/README.md
@@ -42,9 +42,11 @@ use dsh_sdk::DshKafkaConfig;
 use rdkafka::consumer::{Consumer, StreamConsumer};
 use rdkafka::ClientConfig;
 
-fn main() -> Result<(), Box<dyn std::error::Error>>{
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>>{
     // get a rdkafka consumer config for example
     let consumer: StreamConsumer = ClientConfig::new().set_dsh_consumer_config().create()?;
+    Ok(())
 }
 ```
 
@@ -59,15 +61,15 @@ The following features are available in this library and can be enabled/disabled
 
 | **feature** | **default** | **Description** | **Example** |
 | --- |--- | --- | --- |
-| `bootstrap` | &check; | Certificate signing process and fetch datastreams info |  [Kafka](./examples/kafka_example.rs) / [Kafka Proxy](./examples/kafka_proxy.rs) |
+| `bootstrap` | &check; | Certificate signing process and fetch datastreams properties |  [Kafka](./examples/kafka_example.rs) / [Kafka Proxy](./examples/kafka_proxy.rs) |
 | `kafka` |  &check; | Enable `DshKafkaConfig` trait and Config struct to connect to DSH |  [Kafka](./examples/kafka_example.rs) / [Kafka Proxy](./examples/kafka_proxy.rs) |
 | `rdkafka-config` | &check; | Enable `DshKafkaConfig` implementation for RDKafka | [Kafka](./examples/kafka_example.rs) / [Kafka Proxy](./examples/kafka_proxy.rs) |
 | `schema-store` | &cross; | Interact with DSH Schema Store | [Schema Store API](./examples/schema_store_api.rs) |
 | `protocol-token-fetcher` | &cross; | Fetch tokens to use DSH Protocol adapters (MQTT and HTTP) | [Token fetcher](./examples/protocol_token_fetcher.rs) / [with specific claims](./examples/protocol_token_fetcher_specific_claims.rs) |
 | `management-api-token-fetcher` | &cross; | Fetch tokens to use DSH Management API | [ Token fetcher](./examples/management_api_token_fetcher.rs) |
-| `metrics` | &cross; | Enable prometheus metrics including http server | [expose metrics](./examples/expose_metrics.rs) / [custom metrics](./examples/custom_metrics.rs) |
+| `metrics` | &cross; | Enable prometheus metrics including http server | [Expose metrics](./examples/expose_metrics.rs) / [Custom metrics](./examples/custom_metrics.rs) |
 | `graceful-shutdown` | &cross; | Tokio based graceful shutdown handler | [Graceful shutdown](./examples/graceful_shutdown.rs) |
-| `dlq` | &cross; | Dead Letter Queue implementation | [Full implementatione example](./examples/dlq_implementation.rs) |
+| `dlq` | &cross; | Dead Letter Queue implementation | [Full implementation example](./examples/dlq_implementation.rs) |
 
 See the [api documentation](https://docs.rs/dsh_sdk/latest/dsh_sdk/) for more information on how to use these features.
 
@@ -80,7 +82,8 @@ dsh_sdk = { version = "0.5", default-features = false, features = ["management-a
 ```
 
 ## Environment variables
-The SDK checks environment variables to change configuration for  See [ENV_VARIABLES.md](ENV_VARIABLES.md)  which .
+The SDK checks environment variables to change configuration for connnecting to DSH.
+See [ENV_VARIABLES.md](ENV_VARIABLES.md)  which .
 
 ## Examples
 See folder [dsh_sdk/examples](./examples/) for simple examples on how to use the SDK.
