+++ dsh_sdk/src/lib.rs
@@ -1,77 +1,4 @@
-//! # DSH
-//!
-//! Dsh properties struct. Create new to initialize all related components to connect to the DSH kafka clusters and get metadata of your tenant.
-//! - Availablable datastreams info
-//! - Metadata of running container/task
-//! - Certificates for Kafka and DSH
-//!
-//! ## High level API
-//!
-//! The properties struct contains a high level API to interact with the DSH.
-//! This includes generating RDKafka config for creating a consumer/producer and Reqwest config builder for Schema Registry.
-//!
-//! ### Example:
-//! ```
-//! use dsh_sdk::DshKafkaConfig;
-//! use rdkafka::ClientConfig;
-//! use rdkafka::consumer::stream_consumer::StreamConsumer;
-//!
-//! # #[tokio::main]
-//! # async fn main() -> Result<(), Box<dyn std::error::Error>>{
-//! let consumer: StreamConsumer = ClientConfig::new().set_dsh_consumer_config().create()?;
-//! # Ok(())
-//! # }
-//! ```
-//!
-//! ## Low level API
-//! It is also possible to get avaiable metadata or the certificates from the properties struct.
-//!
-//! ### Example:
-//! ```no_run
-//! use dsh_sdk::Dsh;
-
-//! # fn main() -> Result<(), Box<dyn std::error::Error>>{
-//! let dsh = Dsh::get();
-//! // check for write access to topic
-//! let write_access = dsh.datastream().get_stream("scratch.local.local-tenant").expect("Topic not found").write_access();
-//! // get the certificates, for example DSH_KAFKA_CERTIFICATE
-//! let dsh_kafka_certificate = dsh.certificates()?.dsh_kafka_certificate_pem();
-//! #     Ok(())
-//! # }
-//! ```
-//! ## Kafka Proxy / VPN / Local
-//! Read [CONNECT_PROXY_VPN_LOCAL.md](https://github.com/kpn-dsh/dsh-sdk-platform-rs/blob/main/dsh_sdk/CONNECT_PROXY_VPN_LOCAL.md) on how to connect to DSH with Kafka Proxy, VPN or to a local Kafka cluster.
-//!
-//! # Metrics
-//! The metrics module provides a way to expose prometheus metrics. This module is a re-export of the `prometheus` crate. It also contains a function to start a http server to expose the metrics to DSH.
-//!
-//! See [metrics](metrics/index.html) for more information.
-//!
-//! # Graceful shutdown
-//! To implement a graceful shutdown in your service, you can use the `Shutdown` struct. This struct has an implementation based on the best practices example of Tokio.
-//!
-//! This gives you the option to properly handle shutdown in your components/tasks.
-//! It listens for SIGTERM requests and sends out shutdown requests to all shutdown handles.
-//!
-//! See [graceful_shutdown](graceful_shutdown/index.html) for more information.
-//!
-//! # DLQ (Dead Letter Queue)
-//! `OPTIONAL feature: dlq`
-//!
-//! This is an experimental feature and is not yet finalized.
-//!
-//! This implementation only includes pushing messages towards a kafka topic. (Dead or Retry topic)
-//!
-//! ### NOTE:
-//! This implementation does not (and will not) handle any other DLQ related tasks like:
-//!     - Retrying messages
-//!     - Handling messages in DLQ
-//!     - Monitor the DLQ
-//! Above tasks should be handled by a seperate component set up by the user, as these tasks are use case specific and can handle different strategies.
-//!
-//! The DLQ is implemented by running the `Dlq` struct to push messages towards the DLQ topics.
-//! The `ErrorToDlq` trait can be implemented on your defined errors, to be able to send messages towards the DLQ Struct.
-
+#![doc = include_str!("../README.md")]
 #![allow(deprecated)]
 
 // to be kept in v0.6.0
@@ -81,7 +8,9 @@ pub mod certificates;
 pub mod datastream;
 #[cfg(feature = "bootstrap")]
 pub mod dsh;
-pub mod error;
+#[cfg(feature = "bootstrap")]
+mod error;
+
 #[cfg(feature = "management-api-token-fetcher")]
 pub mod management_api;
 pub mod protocol_adapters;
@@ -92,16 +21,15 @@ pub mod schema_store;
 
 #[cfg(feature = "bootstrap")]
 #[doc(inline)]
-pub use dsh::Dsh;
+pub use {dsh::Dsh, error::DshError};
 
 #[cfg(feature = "kafka")]
 #[doc(inline)]
 pub use protocol_adapters::kafka_protocol::DshKafkaConfig;
 
 #[cfg(feature = "management-api-token-fetcher")]
-pub use management_api::token_fetcher::{
-    ManagementApiTokenFetcher, ManagementApiTokenFetcherBuilder,
-};
+#[doc(inline)]
+pub use management_api::{ManagementApiTokenFetcher, ManagementApiTokenFetcherBuilder};
 
 #[doc(inline)]
 pub use utils::Platform;
