+++ dsh_sdk/src/rest_api_token_fetcher.rs
@@ -40,7 +40,7 @@ use std::time::{Duration, Instant};
 use log::debug;
 use serde::Deserialize;
 
-use crate::management_api::error::ManagementTokenError as DshRestTokenError;
+use crate::management_api::ManagementApiTokenError as DshRestTokenError;
 use crate::utils::Platform;
 
 /// Access token of the authentication serveice of DSH.
